+++ src/main/java/dao/factories/GenreDAOSingleton.java
@@ -1,7 +1,7 @@
 package dao.factories;
 
-import dao.GenreDBDAO;
-import dao.GenreMemoryDAO;
+import dao.database.GenreDBDAO;
+import dao.memory.GenreMemoryDAO;
 import dao.api.IGenreDAO;
 
 public class GenreDAOSingleton {
